+++ lib/acfs/model/initialization.rb
@@ -24,9 +24,7 @@ module Acfs::Model
     # @param [ Hash{ Symbol => Object } ] params Attributes to set on resource.
     #
     def initialize(params = {})
-      params.each do |attr, value|
-        self.public_send("#{attr}=", value)
-      end if params
+      self.write_attributes params if params
     end
 
   end